requirements.txt
langchain
langchain-community
langchain_groq
langchain-openai
openai
pyyaml
python-dotenv
streamlit
streamlit-chat